+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added Codecov as an image source to CSP
 - Moved functions out of main file to help with size and DX
 - Explain code block
+- Added "create repo docs" button that replaces "get help from author"
 
 ## [1.1.4]
 - Added github as an image source to CSP